refactor: 프로젝트 파일별 폴더링 구조 개선
