+++ src/components/common/NavigationBar/NavigationBar.style.ts
