+++ src/pages/index.ts
@@ -1,0 +1,1 @@
+export { default as ProductPage } from "./ProductPage/ProductPage";
